Consolidate area equality tests
