+++ pyresample/test/conftest.py
@@ -28,12 +28,75 @@ from pyresample.future.geometry import (
     CoordinateDefinition,
     SwathDefinition,
 )
+from pyresample.geometry import AreaDefinition as LegacyAreaDefinition
+from pyresample.geometry import SwathDefinition as LegacySwathDefinition
 from pyresample.test.utils import create_test_latitude, create_test_longitude
 
 SRC_SWATH_2D_SHAPE = (50, 10)
 SRC_SWATH_1D_SHAPE = (3,)
 SRC_AREA_SHAPE = (50, 10)
 DST_AREA_SHAPE = (80, 85)
+
+
+@pytest.fixture(params=[LegacySwathDefinition, SwathDefinition],
+                ids=["LegacySwathDefinition", "SwathDefinition"])
+def swath_class(request):
+    """Get one of the currently active 'SwathDefinition' classes.
+
+    Currently only includes the legacy 'SwathDefinition' class and the future
+    'SwathDefinition' class in 'pyresample.future.geometry.swath'.
+
+    """
+    return request.param
+
+
+@pytest.fixture
+def create_test_swath(swath_class):
+    """Get a function for creating SwathDefinitions for testing.
+
+    Should be used as a pytest fixture and will automatically run the test
+    function with the legacy SwathDefinition class and the future
+    SwathDefinition class. If tests require a specific class they should
+    NOT use this fixture and instead use the exact class directly.
+
+    """
+    def _create_test_swath(lons, lats):
+        return swath_class(lons, lats)
+    return _create_test_swath
+
+
+@pytest.fixture(params=[LegacyAreaDefinition, AreaDefinition],
+                ids=["LegacyAreaDefinition", "AreaDefinition"])
+def area_class(request):
+    """Get one of the currently active 'AreaDefinition' classes.
+
+    Currently only includes the legacy 'AreaDefinition' class and the future
+    'AreaDefinition' class in 'pyresample.future.geometry.area'.
+
+    """
+    return request.param
+
+
+@pytest.fixture
+def create_test_area(area_class):
+    """Get a function for creating AreaDefinitions for testing.
+
+    Should be used as a pytest fixture and will automatically run the test
+    function with the legacy AreaDefinition class and the future
+    AreaDefinition class. If tests require a specific class they should
+    NOT use this fixture and instead use the exact class directly.
+
+    """
+    def _create_test_area(crs, width, height, area_extent, **kwargs):
+        """Create an AreaDefinition object for testing."""
+        args = (crs, width, height, area_extent)
+        if area_class is LegacyAreaDefinition:
+            attrs = kwargs.pop("attrs", {})
+            area_id = attrs.pop("name", "test_area")
+            args = (area_id, "", "") + args
+        area = area_class(*args, **kwargs)
+        return area
+    return _create_test_area
 
 
 def _euro_lonlats():
