+++ pyresample/test/test_geometry_legacy.py
@@ -81,150 +81,6 @@ class TestBaseDefinition:
 class Test(unittest.TestCase):
     """Unit testing the geometry and geo_filter modules."""
 
-    def test_area_equal(self):
-        """Test areas equality."""
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                               -909968.64000000001,
-                                               1029087.28,
-                                               1490031.3600000001])
-        area_def2 = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                            {'a': '6378144.0',
-                                             'b': '6356759.0',
-                                             'lat_0': '50.00',
-                                             'lat_ts': '50.00',
-                                             'lon_0': '8.00',
-                                             'proj': 'stere'},
-                                            800,
-                                            800,
-                                            [-1370912.72,
-                                                -909968.64000000001,
-                                                1029087.28,
-                                                1490031.3600000001])
-        self.assertFalse(
-            area_def != area_def2, 'area_defs are not equal as expected')
-
-    def test_not_area_equal(self):
-        """Test areas inequality."""
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                               -909968.64000000001,
-                                               1029087.28,
-                                               1490031.3600000001])
-
-        msg_area = geometry.AreaDefinition('msg_full', 'Full globe MSG image 0 degrees',
-                                           'msg_full',
-                                           {'a': '6378169.0',
-                                            'b': '6356584.0',
-                                            'h': '35785831.0',
-                                            'lon_0': '0',
-                                            'proj': 'geos'},
-                                           3712,
-                                           3712,
-                                           [-5568742.4000000004,
-                                               -5568742.4000000004,
-                                               5568742.4000000004,
-                                               5568742.4000000004]
-                                           )
-        self.assertFalse(
-            area_def == msg_area, 'area_defs are not expected to be equal')
-        self.assertFalse(
-            area_def == "area", 'area_defs are not expected to be equal')
-
-    def test_swath_equal_area(self):
-        """Test equality swath area."""
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                               -909968.64000000001,
-                                               1029087.28,
-                                               1490031.3600000001])
-
-        swath_def = geometry.SwathDefinition(*area_def.get_lonlats())
-
-        self.assertFalse(
-            swath_def != area_def, "swath_def and area_def should be equal")
-
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                               -909968.64000000001,
-                                               1029087.28,
-                                               1490031.3600000001])
-
-        self.assertFalse(
-            area_def != swath_def, "swath_def and area_def should be equal")
-
-    def test_swath_not_equal_area(self):
-        """Test inequality swath area."""
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                               -909968.64000000001,
-                                               1029087.28,
-                                               1490031.3600000001])
-
-        lons = np.array([1.2, 1.3, 1.4, 1.5])
-        lats = np.array([65.9, 65.86, 65.82, 65.78])
-        swath_def = geometry.SwathDefinition(lons, lats)
-
-        self.assertFalse(
-            swath_def == area_def, "swath_def and area_def should be different")
-
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)', 'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                               -909968.64000000001,
-                                               1029087.28,
-                                               1490031.3600000001])
-
-        self.assertFalse(
-            area_def == swath_def, "swath_def and area_def should be different")
-
     def test_grid_filter_valid(self):
         """Test valid grid filtering."""
         lons = np.array([-170, -30, 30, 170])
